Split two files back under the 500-LoC cap
CI's file-size job was FAILING the PR. Two files crossed the cap because
several agents appended to each independently across this stack:

  crates/spark-server/src/tui/bench_state.rs   504
  crates/atlas-kernels/src/lib.rs              501

Both are exact piecewise moves with no logic change, following the seams the
files already had.

bench_state.rs -> bench_state_history.rs (480): the two History-pane members,
`load_history` and `elapsed_text`, plus `persist`'s tail. They are the only
members that read ~/.atlas/runs rather than driving a live run, so that is the
natural seam. Declared as a CHILD module via #[path], matching the existing
test modules — a sibling could not see BenchState's private fields, which is
how I first wrote it and how clippy caught it.

lib.rs -> query.rs (487): `available_targets` and `ptx_for_model`, the two
whole-registry lookups. Distinct from `resolve`, which picks ONE target for a
(model_type, hidden_size) and needs a tie-break; these enumerate or
substring-match and cannot fail.

Cap check replicated locally against the workflow's own allow-list: no
violations. fmt clean, clippy exit 0, cargo test --workspace 4841 passed /
0 failed.
